// frontend/src/api/api.ts
export interface SignUpDto {
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'phone': string;
    /**
     * 
     * @type {number}
     * @memberof SignUpDto
     */
    'age': number;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'password': string;
}
/**
 * 
 * @export
 * @interface SignUpResponseDto
 */
export interface SignUpResponseDto {
    /**
     * 
     * @type {User}
     * @memberof SignUpResponseDto
     */
    'user': User;
    /**
     * 
     * @type {string}
     * @memberof SignUpResponseDto
     */
    'accessToken': string;
}
/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'phone': string;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    'age': number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'password': string;
}